requirements.txt
streamlit
tensorflow == 2.18.0
requests
Pillow
numpy
